+++ CHANGELOG.md
@@ -23,6 +23,7 @@
     `restrict_user_defined_variables` settings.
   * Projects can now be created and edited with
     `allow_merge_on_skipped_pipeline` settings.
+  * Projects can now be edited with `ci_forward_deployment_enabled` settings.
 
 # v0.1308.0